qr_secret/requirements.txt
asgiref==3.11.1
Django==3.2.16
pillow==11.3.0
pytz==2026.1.post1
qrcode==8.2
sqlparse==0.5.5
typing_extensions==4.15.0
dj-database-url==0.5.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
whitenoise==6.7.0
